requirements.txt
python-dotenv
google-genai
google-adk
rich